fix(ci): split oversized ha standby startup test
